Refactor Experience packages
