+++ src/main/java/edu/oregonstate/cs467/travelplanner/experience/persistence/ExperienceSearchDao.java
@@ -1,7 +1,9 @@
-package edu.oregonstate.cs467.travelplanner.experience.persistence.search;
+package edu.oregonstate.cs467.travelplanner.experience.persistence;
 
 import edu.oregonstate.cs467.travelplanner.experience.model.Experience;
-import edu.oregonstate.cs467.travelplanner.experience.persistence.search.ExperienceSearchParams.ExperienceSearchSort;
+import edu.oregonstate.cs467.travelplanner.experience.service.dto.ExperienceSearchParams;
+import edu.oregonstate.cs467.travelplanner.experience.service.dto.ExperienceSearchParams.ExperienceSearchSort;
+import edu.oregonstate.cs467.travelplanner.experience.service.dto.ExperienceSearchResult;
 import jakarta.validation.Valid;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.JdbcClient;
