+++ src/main/java/edu/oregonstate/cs467/travelplanner/experience/service/dto/CreateUpdateExperienceDto.java
@@ -1,4 +1,4 @@
-package edu.oregonstate.cs467.travelplanner.experience.dto;
+package edu.oregonstate.cs467.travelplanner.experience.service.dto;
 
 import edu.oregonstate.cs467.travelplanner.experience.model.Experience;
 import edu.oregonstate.cs467.travelplanner.util.validation.NotBlankNull;
