+++ src/main/java/edu/oregonstate/cs467/travelplanner/experience/service/dto/ExperienceSearchResult.java
@@ -1,4 +1,4 @@
-package edu.oregonstate.cs467.travelplanner.experience.persistence.search;
+package edu.oregonstate.cs467.travelplanner.experience.service.dto;
 
 import edu.oregonstate.cs467.travelplanner.experience.model.Experience;
