+++ src/main/java/edu/oregonstate/cs467/travelplanner/experience/web/ExperienceWebController.java
@@ -1,6 +1,6 @@
-package edu.oregonstate.cs467.travelplanner.experience.controller;
+package edu.oregonstate.cs467.travelplanner.experience.web;
 
-import edu.oregonstate.cs467.travelplanner.experience.dto.CreateUpdateExperienceDto;
+import edu.oregonstate.cs467.travelplanner.experience.service.dto.CreateUpdateExperienceDto;
 import edu.oregonstate.cs467.travelplanner.experience.model.Experience;
 import edu.oregonstate.cs467.travelplanner.experience.service.ExperienceService;
 import edu.oregonstate.cs467.travelplanner.trip.model.Trip;
