+++ src/test/java/edu/oregonstate/cs467/travelplanner/experience/service/dto/ExperienceSearchParamsTests.java
@@ -1,8 +1,8 @@
-package edu.oregonstate.cs467.travelplanner.experience.persistence.search;
+package edu.oregonstate.cs467.travelplanner.experience.service.dto;
 
 import edu.oregonstate.cs467.travelplanner.AbstractBaseTest;
-import edu.oregonstate.cs467.travelplanner.experience.persistence.search.ExperienceSearchParams.ExperienceSearchLocationParams;
-import edu.oregonstate.cs467.travelplanner.experience.persistence.search.ExperienceSearchParams.ExperienceSearchSort;
+import edu.oregonstate.cs467.travelplanner.experience.service.dto.ExperienceSearchParams.ExperienceSearchLocationParams;
+import edu.oregonstate.cs467.travelplanner.experience.service.dto.ExperienceSearchParams.ExperienceSearchSort;
 import jakarta.validation.Validator;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
